Move three functions from notebooks/util.py to main package
