+++ text_extensions_for_pandas/test_util.py
@@ -13,65 +13,12 @@
 #  limitations under the License.
 #
 
-import textwrap
 import unittest
 
-import pandas as pd
-import regex
-import spacy
-
-_SPACY_LANGUAGE_MODEL = spacy.load("en_core_web_sm")
-
-from text_extensions_for_pandas.util import TestBase, pretty_print_html
-from text_extensions_for_pandas.io import make_tokens_and_features
-
-
-_TEST_TEXT = "Item's for < $100 & change"
-_TEST_TOKS = make_tokens_and_features(_TEST_TEXT, _SPACY_LANGUAGE_MODEL)
-
+from text_extensions_for_pandas.util import TestBase
 
 class UtilTest(TestBase):
-    def test_pretty_print_html(self):
-        self.maxDiff = None
-        html = pretty_print_html(_TEST_TOKS["token_span"].values, True)
-        suffix = html[-809:]
-        # print(f"[[[{suffix}]]]")
-        self.assertEqual(
-            suffix,
-            """\
-    <td>20</td>
-      <td>26</td>
-      <td>7</td>
-      <td>8</td>
-      <td>change</td>
-    </tr>
-  </tbody>
-</table>
-            </div>
-            <div id="text"
-             style="float:right; background-color:#F5F5F5; border: 1px solid #E0E0E0; width: 60%;">
-                <div style="float:center; padding:10px">
-                    <p style="font-family:monospace">
-                        <span style="background-color:yellow">Item&#39;s</span> <span style="background-color:yellow">for</span> <span style="background-color:yellow">&lt;</span> <span style="background-color:yellow"><span>&#36;</span>100</span> <span style="background-color:yellow">&amp;</span> <span style="background-color:yellow">change
-                    </p>
-                </div>
-            </div>
-        </div>
-        """)
+    # Nothing to test at the moment, besides the "meta" test that you can create
+    # a subclass of util.TestBase
+    pass
 
-        html = pretty_print_html(_TEST_TOKS["token_span"].values, False)
-        suffix = html[-809:]
-        # print(f"[[[{suffix}]]]")
-        self.assertEqual(
-            suffix,
-            """\
-
-        <div id="text"
-         style="float:right; background-color:#F5F5F5; border: 1px solid #E0E0E0; width: 100%;">
-            <div style="float:center; padding:10px">
-                <p style="font-family:monospace">
-                    <span style="background-color:yellow">Item&#39;s</span> <span style="background-color:yellow">for</span> <span style="background-color:yellow">&lt;</span> <span style="background-color:yellow"><span>&#36;</span>100</span> <span style="background-color:yellow">&amp;</span> <span style="background-color:yellow">change
-                </p>
-            </div>
-        </div>
-        """)
